Player Class 수정
